apps/daemon/requirements.txt
uiautomator2==3.3.1
fastapi==0.111.0
uvicorn==0.30.0
websockets==12.0
python-dotenv==1.0.1
pillow==10.3.0
httpx==0.27.0
aiofiles==23.2.1
